+++ ricaun.Revit.UI/Utils/StackTraceUtils.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 
@@ -6,6 +7,26 @@ namespace ricaun.Revit.UI.Utils
 {
     internal static class StackTraceUtils
     {
+        /// <summary>
+        /// Check if assembly is repacked.
+        /// </summary>
+        internal static bool IsAssemblyRepack { get; } = GetAssemblyRepack();
+
+        /// <summary>
+        /// The current assembly does not have any resource and name starts with 'ricaun.Revit.UI' namespace.
+        /// </summary>
+        /// <returns></returns>
+        private static bool GetAssemblyRepack()
+        {
+            var assembly = Assembly.GetExecutingAssembly();
+
+            var isAssemblyName = assembly.GetName().Name.StartsWith(typeof(AppLoaderAttribute).Namespace);
+            if (!isAssemblyName) 
+                return true;
+
+            return assembly.GetManifestResourceNames().Length != 0;
+        }
+
         /// <summary>
         /// GetCallingAssembly using StackTrace
         /// </summary>
@@ -16,6 +37,12 @@ namespace ricaun.Revit.UI.Utils
             var callingAssembly = Assembly.GetCallingAssembly();
             var executingAssembly = Assembly.GetExecutingAssembly();
             if (callingAssembly != executingAssembly) { return callingAssembly; }
+
+            if (IsAssemblyRepack)
+            {
+                //Debug.WriteLine($"Assembly is repacked, return {callingAssembly}.");
+                return callingAssembly;
+            }
 
             const int skipFrames = 1; // Skip 'StackTraceUtils' method. 
             var stackFrames = new StackTrace(skipFrames).GetFrames();
@@ -31,7 +58,6 @@ namespace ricaun.Revit.UI.Utils
                     break;
                 }
             }
-
             return callingAssembly;
         }