+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] / 2024-12-27 - 2025-02-18
+### Features
+- Support `RevitApplication` static class.
+- Support `RevitApplicationPreview` static class.
+### Updated
+- Add `RevitApplication` with static `UIApplication`, `UIControlledApplication` and `IsInAddInContext`.
+- Add `RevitApplicationPreview` with static `IsPreviewRelease`.
+- Update `StackTraceUtils` to support repack using `IsAssemblyRepack`.
+### Tests
+- Test `RevitApplication` static class.
+
 ## [0.7.1] / 2024-11-11 - 2024-11-19
 ### Features
 - Support `tiff` files for icon.
@@ -378,6 +389,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.8.0]: ../../compare/0.7.1...0.8.0
 [0.7.1]: ../../compare/0.7.0...0.7.1
 [0.7.0]: ../../compare/0.6.2...0.7.0
 [0.6.2]: ../../compare/0.6.1...0.6.2